Made Min/MaxValue more general

+++ distributions/univariate/discrete/BetaBinomialRand.h
@@ -17,8 +17,8 @@ public:
     BetaBinomialRand(int number, double shape1, double shape2);
     std::string name() const override;
     SUPPORT_TYPE supportType() const override { return FINITE_T; }
-    double MinValue() const override { return 0; }
-    double MaxValue() const override { return n; }
+    int MinValue() const override { return 0; }
+    int MaxValue() const override { return n; }
 
     void setParameters(int number, double shape1, double shape2);
     inline int getNumber() const { return n; }
