+++ distributions/univariate/discrete/BinomialRand.h
@@ -36,8 +36,8 @@ public:
     BinomialRand(int number, double probability);
     std::string name() const override;
     SUPPORT_TYPE supportType() const override { return FINITE_T; }
-    double MinValue() const override { return 0; }
-    double MaxValue() const override { return n; }
+    int MinValue() const override { return 0; }
+    int MaxValue() const override { return n; }
 
 private:
     void setGeneratorConstants();
