+++ distributions/univariate/discrete/RademacherRand.h
@@ -13,8 +13,8 @@ public:
     RademacherRand();
     std::string name() const override;
     SUPPORT_TYPE supportType() const override { return FINITE_T; }
-    double MinValue() const override { return -1; }
-    double MaxValue() const override { return 1; }
+    int MinValue() const override { return -1; }
+    int MaxValue() const override { return 1; }
 
     double P(int k) const override;
     double F(int k) const override;
